feat(boardDelete): 유저 정보를 확인하여 id를 점검하는 로직 추가
- 유저정보를 json 파일로 전달
- user 기능과 board 기능 통합 후 전달받는 json 형식 변경

+++ src/main/java/com/example/myreform/service/board/BoardServiceImpl.java
@@ -17,7 +17,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -87,11 +86,16 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public String delete(Long boardId) {
+    public String delete(Long boardId, User user) {
         if (!boardRepository.existsById(boardId) || findById(boardId).getStatus() == 0) {
             return "해당 게시글이 없습니다.";
         }
         Board board = findById(boardId);
+        if (!board.getUser().getUserId().equals(user.getUserId())) {
+            System.out.println(board.getUser().getUserId());
+            System.out.println(user.getUserId());
+            return "게시물 삭제 권한이 없습니다.";
+        }
         board.delete(); // status만 수정
         List<BoardImage> postImages = boardImageRepository.findAllByBoardId(boardId);
         deleteBoardImages(postImages);
